refactoring: [Task list page layout refactoring]change folder structure
