+++ src/pages/index.js
@@ -1,6 +1,6 @@
 import CartPage from './cart-page';
 import HomePage from './home-page';
-import TasksTable from './tasksTable';
+import TasksTable from './tasksTable/tasksTable';
 import Check from './check/check';
 import CrossCheckSessions from './crossCheckSessions';
 import ReviewsListPage from './reviewsListPage';
